chore: move db
