+++ apps/backend/.env.example
@@ -1,5 +1,5 @@
 # prod | dev | local
-STAGE=local 
+STAGE=local
 
 # For Mapbox API. Can be an empty string
 MAPBOX_ACCESS_TOKEN=pk.abc123
