--- apps/backend/src/db/schema/auth/index.ts
@@ -1,3 +1,0 @@
-export * from './account'
-export * from './session'
-export * from './user'
