--- apps/backend/src/db/schema/index.ts
@@ -1,3 +1,0 @@
-export * from './auth'
-export * from './schedule'
-export * from './subscription'
