+++ packages/db/.env.example
@@ -1,0 +1,1 @@
+DB_URL="postgres://postgres:postgres@localhost:5432/antalmanac"
