+++ packages/db/Dockerfile
@@ -1,0 +1,1 @@
+FROM postgres:16.8-alpine3.21
