+++ packages/db/src/index.ts
@@ -1,0 +1,4 @@
+import 'dotenv/config';
+import { drizzle } from 'drizzle-orm/node-postgres';
+
+export const db = drizzle(process.env.DATABASE_URL as string);
