+++ packages/db/src/schema/auth/account.ts
@@ -2,15 +2,11 @@ import { primaryKey, pgTable, text, pgEnum } from 'drizzle-orm/pg-core';
 
 import { users } from './user';
 
-
 const accountTypes = ['GOOGLE', 'GUEST'] as const;
 
-export const accountTypeEnum = pgEnum(
-    'account_type',
-    accountTypes
-);
+export const accountTypeEnum = pgEnum('account_type', accountTypes);
 
-export type AccountType = typeof accountTypes[number];
+export type AccountType = (typeof accountTypes)[number];
 
 // Each user can have multiple accounts, each account is associated with a provider.
 // A user without an account is a username-only user.
@@ -27,9 +23,7 @@ export const accounts = pgTable(
 
         providerAccountId: text('provider_account_id').notNull(),
     },
-    (table) => ([
-        primaryKey({columns: [table.userId, table.accountType], }),
-    ])
+    (table) => [primaryKey({ columns: [table.userId, table.accountType] })]
 );
 
 export type Account = typeof accounts.$inferSelect;
