+++ packages/db/src/schema/auth/index.ts
@@ -1,0 +1,3 @@
+export * from './account';
+export * from './session';
+export * from './user';
