+++ packages/db/src/schema/schedule/index.ts
