+++ packages/db/tsconfig.json
@@ -1,0 +1,16 @@
+{
+    "compilerOptions": {
+        "target": "esnext",
+        "module": "esnext",
+        "moduleResolution": "node",
+        "esModuleInterop": true,
+        "experimentalDecorators": true,
+        "strict": true,
+        "skipLibCheck": true,
+        "baseUrl": ".",
+        "outDir": "dist",
+        "resolveJsonModule": true
+    },
+    "include": ["drizzle.config.ts"],
+    "exclude": ["node_modules", "dist"]
+}
