+++ tsconfig.json
@@ -1,3 +1,6 @@
 {
     // Empty tsconfig for VSCode users
+    "compilerOptions": {
+        "allowSyntheticDefaultImports": true
+    }
 }
